requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai==0.6.0
Werkzeug==3.0.0
Jinja2==3.1.2
